Add files via upload
Takes input shapefile of polygons, and slices them in 1 of the 8 main cardinal directions (N,NE,E,SE,S,SW,W,NW). Script explains what the shapefile database needs to include. 
Non-routine code at the start of the script for a special case I was working on. Some editing needs to be done to remove this extraneous code for general use.